requirements.txt
# Dependencies for TRUST MCNet
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
PyYAML>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
